aider: The Serilog extension now loads the UserSerilogRequestLogging.
# Aider chat conversation:

USER: can you make sure the serilog extension loads the UserSerilogRequestLogging?

ASSISTANT: Sure, I can help with that. We need to modify the `src/Infrastructure/SerilogExtensions.cs` file to ensure that the `UseSerilogRequestLogging` method is called when configuring Serilog.

Here is the change:

```csharp
src/Infrastructure/SerilogExtensions.cs
<<<<<<< ORIGINAL
        // code to register
        //app.UseSerilogRequestLogging();
=======
        // code to register
        builder.Host.UseSerilogRequestLogging();
>>>>>>> UPDATED
```

This change will ensure that the `UseSerilogRequestLogging` method is called when the `UseSerilog` extension method is used to configure Serilog. This will enable request logging in the application.