aider: Extend `WebApplication` with `UseSerilogRequestLogging` and set custom options for `EnrichDiagnosticContext` and `GetLevel` in `Program.cs`.
# Aider chat conversation:

USER: that's not quite right, the UseSerilogRequestLogging needs to extend the `WebApplication`

ASSISTANT: I see, you want to extend the `WebApplication` with `UseSerilogRequestLogging`. In that case, we need to modify the `src/Program.cs` file. Here is the change:

```csharp
src/Program.cs
<<<<<<< ORIGINAL
app.UseSerilogRequestLogging();
=======
app.UseSerilogRequestLogging(options =>
{
    options.EnrichDiagnosticContext = LogHelper.EnrichFromRequest;
    options.GetLevel = LogHelper.CustomGetLevel;
});
>>>>>>> UPDATED
```

This change will ensure that the `UseSerilogRequestLogging` method is called when the `WebApplication` is built. The `EnrichDiagnosticContext` and `GetLevel` options are set to custom methods defined in `LogHelper`. If these methods are not yet defined, you will need to implement them.